fix(range-selector): skip validation during the initialization phase
